% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanVeg.R
\name{getMeanVeg}
\alias{getMeanVeg}
\title{getMeanVeg Function}
\usage{
getMeanVeg(id, choosen_cen)
}
\arguments{
\item{id}{Individual's ID}

\item{census}{A group of census data that is probably subset beforehand if
you are looking for a particular set of data (i.e. by season, year)}
}
\description{
This function allows you to get a weighted mean of the vegetation in
certain census coordinate points.
}
\keyword{vegetation}
