% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopDensity.R
\name{getPopDensity}
\alias{getPopDensity}
\title{getPopDensity Function}
\usage{
getPopDensity(census)
}
\arguments{
\item{census}{Takes in a data frame of census data that is probably subset beforehand if
you are looking for a particular set of data (i.e. by season, year)}
}
\description{
This function allows you to get an indice of local population density,
based off of where the individual was seen in census data.
}
\keyword{density}
\keyword{population}
